+++ CHANGELOG.md
@@ -79,3 +79,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - **Tag Transfer:** Fixed tag transfer functionality.
+- **Debt:** Fixed debt allocation functionality.